+++ src/components/AccountRow.tsx
@@ -6,7 +6,7 @@ import { Box, Button, IconButton, SxProps, Typography } from "@mui/material";
 import { Link, Navigate, useMatch } from "react-router-dom";
 
 import EditAccountPopper from "components/EditAccountPopper";
-import SolidSelectable from "components/SolidSelectable";
+import SolidSelectable from "components/solid/SolidSelectable";
 import { useAccount, useTransactionsByAccountID } from "data/Hooks";
 // import { Account } from "models/Budget";
 import { formatCurrencyCents } from "Utils";
