+++ src/components/transactions/TransactionRow.tsx
@@ -31,40 +31,40 @@ const Item = ({
     );
 };
 
-const HoverableTypography = ({
-    isSelected,
-    noWrap,
-    children,
-    sx,
-}: {
-    isSelected: boolean;
-    noWrap?: boolean;
-    children: React.ReactNode;
-    sx?: SxProps;
-}) => {
-    return (
-        <Typography
-            noWrap={noWrap}
-            sx={{
-                width: 1,
-                // minWidth: "1rem",
-                minHeight: "1.5rem",
-                textAlign: "left",
-                ":hover": isSelected
-                    ? {
-                          outlineColor: (theme) => theme.palette.primary.dark,
-                          outlineStyle: "solid",
-                          outlineWidth: "0.1rem",
-                          borderRadius: "0.18rem",
-                      }
-                    : undefined,
-                ...sx,
-            }}
-        >
-            {children}
-        </Typography>
-    );
-};
+// const HoverableTypography = ({
+//     isSelected,
+//     noWrap,
+//     children,
+//     sx,
+// }: {
+//     isSelected: boolean;
+//     noWrap?: boolean;
+//     children: React.ReactNode;
+//     sx?: SxProps;
+// }) => {
+//     return (
+//         <Typography
+//             noWrap={noWrap}
+//             sx={{
+//                 width: 1,
+//                 // minWidth: "1rem",
+//                 minHeight: "1.5rem",
+//                 textAlign: "left",
+//                 ":hover": isSelected
+//                     ? {
+//                           outlineColor: (theme) => theme.palette.primary.dark,
+//                           outlineStyle: "solid",
+//                           outlineWidth: "0.1rem",
+//                           borderRadius: "0.18rem",
+//                       }
+//                     : undefined,
+//                 ...sx,
+//             }}
+//         >
+//             {children}
+//         </Typography>
+//     );
+// };
 
 const TransactionRow = ({
     isSelected,
@@ -87,10 +87,10 @@ const TransactionRow = ({
     let columnIndex = 0;
 
     const bgColor = (() => {
-        if (isEditing) {
-            return index % 2 === 0 ? "red" : "green";
-        }
-        if (isSelected) {
+        // if (isEditing) {
+        //     return index % 2 === 0 ? "red" : "green";
+        // }
+        if (isSelected || isEditing) {
             return index % 2 === 0 ? "primary.light" : "primary.lighter";
         }
         return index % 2 === 0 ? "neutral.light" : "white";
@@ -124,9 +124,13 @@ const TransactionRow = ({
                     width: columnRatios[columnIndex++],
                 }}
             >
-                <Typography noWrap>
-                    {dayjs(transaction.timestamp).format("YYYY-MM-DD")}
-                </Typography>
+                {isEditing ? (
+                    <Typography></Typography>
+                ) : (
+                    <Typography noWrap>
+                        {dayjs(transaction.timestamp).format("YYYY-MM-DD")}
+                    </Typography>
+                )}
             </Item>
             <Item
                 sx={{
